refactor: 로또 비교 기능 등수 계산 클래스로 이동
