+++ src/test/java/lotto/temp/StaticsTest.java
@@ -15,19 +15,7 @@ class StaticsTest {
 
     @BeforeEach
     void setUp() {
-        statics = new Statics(new CommonIo());
-    }
-
-    @Test
-    @DisplayName("구입한 로또와 당첨번호가 몇 개 일치하는지 확인하는 테스트")
-    void compareLottos(){
-        assertThat(statics.compareLottos(
-                List.of(Lotto.createUserLotto(List.of(1,2,3,4,5,6)),
-                        Lotto.createUserLotto(List.of(1,2,3,4,5,7)),
-                        Lotto.createUserLotto(List.of(1,2,3,4,7,8)),
-                        Lotto.createUserLotto(List.of(1,2,3,7,8,9)),
-                        Lotto.createUserLotto(List.of(1,2,7,8,9,10))),
-                List.of(1,2,3,4,5,6))).containsExactly(6,5,4,3,2);
+        statics = new Statics();
     }
 
     @Test
